+++ go-ethereum/core/types/gen_withdrawal_rlp.go
@@ -8,10 +8,10 @@ import "io"
 func (obj *Withdrawal) EncodeRLP(_w io.Writer) error {
 	w := rlp.NewEncoderBuffer(_w)
 	_tmp0 := w.List()
-	w.WriteUint64(uint64(obj.Index))
-	w.WriteUint64(uint64(obj.Validator))
+	w.WriteUint64(obj.Index)
+	w.WriteUint64(obj.Validator)
 	w.WriteBytes(obj.Address[:])
-	w.WriteUint64(uint64(obj.Amount))
+	w.WriteUint64(obj.Amount)
 	w.ListEnd(_tmp0)
 	return w.Flush()
 }
